requirements.txt
Django
discord.pu